requirements.txt
requests>=2.28.0
pyyaml>=6.0
aiohttp>=3.8.0
tenacity>=8.0.0
rich>=13.0.0
